+++ database/migrations/2020_09_27_000173_add_relationship_fields_to_opportunities_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToOpportunitiesTable extends Migration
     public function up()
     {
         Schema::table('opportunities', function (Blueprint $table) {
-            $table->unsignedInteger('lead_id')->nullable();
+            $table->unsignedBigInteger('lead_id')->nullable();
             $table->foreign('lead_id', 'lead_fk_2172516')->references('id')->on('leads');
         });
     }
